requirements.txt
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
APScheduler==3.11.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.1
attrs==24.2.0
bcrypt==4.1.3
bidict==0.23.1
blinker==1.9.0
cachetools==5.3.3
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.4.4
choreographer==1.2.1
click==8.3.1
colorama==0.4.6
comm==0.2.3
cryptography==44.0.1
darkdetect==0.8.0
debugpy==1.8.19
decorator==5.2.1
Deprecated==1.3.1
dnspython==2.8.0
duckdb==1.4.3
email-validator==2.3.0
executing==2.2.1
fastjsonschema==2.21.2
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Cors==6.0.2
Flask-Limiter==3.7.0
Flask-Login==0.6.3
flask-restx==1.3.2
Flask-SocketIO==5.6.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.3.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx[http2]==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
importlib-resources==6.5.2
iniconfig==2.0
ipykernel==6.29.5
ipython==9.9.0
ipython-pygments-lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2023.12.1
jupyter-client==8.8.0
jupyter-core==5.9.0
kaleido==1.2.0
limits==3.13.0
llvmlite==0.46.0b1
logistro==2.0.1
logzero==1.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.2
matplotlib-inline==0.2.1
mcp==1.26.0
mdurl==0.1.2
py_lets_be_rational==1.0.1
py_vollib==1.0.1
narwhals==2.15.0
nbformat==5.10.4
nest-asyncio==1.6.0
numba==0.63.1
numpy==2.3.5
openalgo==1.0.45
ordered-set==4.1.0
orjson==3.11.5
packaging==24.1
pandas==2.3.3
pyarrow==22.0.0
fastparquet==2025.12.0
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
plotly==6.5.2
pluggy==1.0
prompt-toolkit==3.0.52
protobuf==6.33.4
psutil==7.2.1
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==2.22
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
Pygments==2.19.2
PyJWT==2.10.1
pyngrok==7.5.0
pyotp==2.9.0
pytest==9.0.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.0
python-multipart==0.0.22
python-socketio==5.16.0
python-telegram-bot==22.6
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
qrcode==8.2
referencing==0.37.0
requests==2.32.5
rich==13.7.1
rpds-py==0.30.0
scipy==1.17.0
setuptools==80.10.1
simple-websocket==1.1.0
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.46
sse-starlette==2.4.1
stack-data==0.6.3
starlette==0.52.1
tornado==6.5.4
traitlets==5.14.3
typing-extensions>=4.12.2
typing-inspection>=0.4.1
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
wcwidth==0.3.5
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.5
wheel==0.46.3
wrapt==1.16.0
wsproto==1.3.2
WTForms==3.2.1
zipp==3.23.0
zmq==0.0.0

# US Market Brokers
ibapi>=5.19.2
alpaca-trade-api>=3.0.0
polygon-api-client>=1.13.0

# US Market Brokers
ibapi>=5.19.2
alpaca-trade-api>=3.0.0
polygon-api-client>=1.13.0
kalshi-python-sync>=1.0.0
